+++ R/ck_params_cnts.R
@@ -4,7 +4,8 @@
 #' to perturb count variables.
 #'
 #' @inheritParams ck_params_nums
-#' @param ptab an object created with [ptable::pt_create_pTable()]
+#' @param ptab an object created with [ptable::pt_create_pParams()] or
+#' [ptable::pt_create_pTable()]
 #' @return an object suitable as input to method `$params_cnts_set()` for the perturbation
 #' of counts and frequencies.
 #' @export
@@ -15,6 +16,13 @@
 #' is available from the respective help-pages of these functions.
 #' @inherit cellkey_pkg examples
 ck_params_cnts <- function(ptab, path = NULL) {
+  if (inherits(ptab, "ptable_params")) {
+    if (ptab@table != "cnts") {
+      e <- "input from `ptable::pt_create_pParams` not suitable for count variables."
+      stop(e, call. = FALSE)
+    }
+    ptab <- ptable::pt_create_pTable(params = ptab)
+  }
   out <- list(
     params = list(ptable = .chk_ptab(ptab, type = "cnts")),
     type = "cnts")